Add conversation entity
