+++ custom_components/extended_openai_conversation/conversation.py
@@ -1,0 +1,557 @@
+from __future__ import annotations
+
+import json
+import logging
+from typing import Literal
+
+from openai import AsyncAzureOpenAI, AsyncOpenAI
+from openai._exceptions import AuthenticationError, OpenAIError
+from openai.types.chat.chat_completion import (
+    ChatCompletion,
+    ChatCompletionMessage,
+    Choice,
+)
+import yaml
+
+from homeassistant.components import conversation
+from homeassistant.components.conversation import (
+    ChatLog,
+    ConversationEntity,
+    ConversationEntityFeature,
+    ConversationInput,
+    ConversationResult,
+    async_get_chat_log,
+)
+from homeassistant.components.homeassistant.exposed_entities import async_should_expose
+from homeassistant.config_entries import ConfigEntry
+from homeassistant.const import ATTR_NAME, CONF_API_KEY, MATCH_ALL
+from homeassistant.core import HomeAssistant
+from homeassistant.exceptions import (
+    ConfigEntryNotReady,
+    HomeAssistantError,
+    TemplateError,
+)
+from homeassistant.helpers import (
+    config_validation as cv,
+    device_registry as dr,
+    entity_registry as er,
+    intent,
+    template,
+)
+from homeassistant.helpers.chat_session import async_get_chat_session
+from homeassistant.helpers.entity_platform import AddEntitiesCallback
+from homeassistant.helpers.httpx_client import get_async_client
+from homeassistant.helpers.typing import ConfigType
+from homeassistant.util import ulid
+
+from .const import (
+    CONF_API_VERSION,
+    CONF_ATTACH_USERNAME,
+    CONF_BASE_URL,
+    CONF_CHAT_MODEL,
+    CONF_CONTEXT_THRESHOLD,
+    CONF_CONTEXT_TRUNCATE_STRATEGY,
+    CONF_FUNCTIONS,
+    CONF_MAX_FUNCTION_CALLS_PER_CONVERSATION,
+    CONF_MAX_TOKENS,
+    CONF_ORGANIZATION,
+    CONF_PROMPT,
+    CONF_SKIP_AUTHENTICATION,
+    CONF_TEMPERATURE,
+    CONF_TOP_P,
+    CONF_USE_TOOLS,
+    DEFAULT_ATTACH_USERNAME,
+    DEFAULT_CHAT_MODEL,
+    DEFAULT_CONF_FUNCTIONS,
+    DEFAULT_CONTEXT_THRESHOLD,
+    DEFAULT_CONTEXT_TRUNCATE_STRATEGY,
+    DEFAULT_MAX_FUNCTION_CALLS_PER_CONVERSATION,
+    DEFAULT_MAX_TOKENS,
+    DEFAULT_PROMPT,
+    DEFAULT_SKIP_AUTHENTICATION,
+    DEFAULT_TEMPERATURE,
+    DEFAULT_TOP_P,
+    DEFAULT_USE_TOOLS,
+    DOMAIN,
+    EVENT_CONVERSATION_FINISHED,
+)
+from .exceptions import (
+    FunctionLoadFailed,
+    FunctionNotFound,
+    InvalidFunction,
+    ParseArgumentsFailed,
+    TokenLengthExceededError,
+)
+from .helpers import get_function_executor, is_azure, validate_authentication
+from .services import async_setup_services
+
+DATA_AGENT = "agent"
+_LOGGER = logging.getLogger(__name__)
+
+
+async def async_setup_entry(
+    hass: HomeAssistant,
+    config_entry: ConfigEntry,
+    async_add_entities: AddEntitiesCallback,
+) -> None:
+    """Set up the OpenAI Conversation entities."""
+    _LOGGER.info(
+        "Extended OpenAI Conversation async_setup_entry called: %s", config_entry.data
+    )
+    try:
+        await validate_authentication(
+            hass=hass,
+            api_key=config_entry.data[CONF_API_KEY],
+            base_url=config_entry.data.get(CONF_BASE_URL),
+            api_version=config_entry.data.get(CONF_API_VERSION),
+            organization=config_entry.data.get(CONF_ORGANIZATION),
+            skip_authentication=config_entry.data.get(
+                CONF_SKIP_AUTHENTICATION, DEFAULT_SKIP_AUTHENTICATION
+            ),
+        )
+    except AuthenticationError as err:
+        _LOGGER.error("Invalid API key: %s", err)
+        raise ConfigEntryNotReady("Invalid API key") from err
+    except OpenAIError as err:
+        raise ConfigEntryNotReady(err) from err
+
+    agent = ExtendedOpenAIAgent(hass, config_entry)
+
+    data = hass.data.setdefault(DOMAIN, {}).setdefault(config_entry.entry_id, {})
+    data[CONF_API_KEY] = config_entry.data[CONF_API_KEY]
+    data[DATA_AGENT] = agent
+
+    async_add_entities([agent])
+    # conversation.async_set_agent(hass, config_entry, agent)
+
+
+class ExtendedOpenAIAgent(ConversationEntity, conversation.AbstractConversationAgent):
+    """OpenAI conversation agent."""
+
+    _attr_has_entity_name = True
+    _attr_name = None
+    _attr_supported_features = ConversationEntityFeature.CONTROL
+
+    def __init__(self, hass: HomeAssistant, entry: ConfigEntry) -> None:
+        """Initialize the agent."""
+        self.hass = hass
+        self.entry = entry
+        self.history: dict[str, list[dict]] = {}
+
+        self._attr_unique_id = entry.entry_id
+        self._attr_device_info = dr.DeviceInfo(
+            identifiers={(DOMAIN, entry.entry_id)},
+            name=entry.title,
+            manufacturer="OpenAI",
+            model=self.entry.options.get(CONF_CHAT_MODEL, DEFAULT_CHAT_MODEL),
+            entry_type=dr.DeviceEntryType.SERVICE,
+        )
+
+        base_url = entry.data.get(CONF_BASE_URL)
+        if is_azure(base_url):
+            self.client = AsyncAzureOpenAI(
+                api_key=entry.data[CONF_API_KEY],
+                azure_endpoint=base_url,
+                api_version=entry.data.get(CONF_API_VERSION),
+                organization=entry.data.get(CONF_ORGANIZATION),
+                http_client=get_async_client(hass),
+            )
+        else:
+            self.client = AsyncOpenAI(
+                api_key=entry.data[CONF_API_KEY],
+                base_url=base_url,
+                organization=entry.data.get(CONF_ORGANIZATION),
+                http_client=get_async_client(hass),
+            )
+        # Cache current platform data which gets added to each request (caching done by library)
+        _ = hass.async_add_executor_job(self.client.platform_headers)
+
+    @property
+    def supported_languages(self) -> list[str] | Literal["*"]:
+        """Return a list of supported languages."""
+        return MATCH_ALL
+
+    async def async_process(self, user_input: ConversationInput) -> ConversationResult:
+        """Process a sentence."""
+        with (
+            async_get_chat_session(self.hass, user_input.conversation_id) as session,
+            async_get_chat_log(self.hass, session, user_input) as chat_log,
+        ):
+            return await self._async_handle_message(user_input, chat_log)
+
+    async def _async_handle_message(
+        self,
+        user_input: ConversationInput,
+        chat_log: ChatLog,
+    ) -> ConversationResult:
+        """Call the API."""
+        exposed_entities = self.get_exposed_entities()
+
+        conversation_id = chat_log.conversation_id
+        if conversation_id in self.history:
+            messages = self.history[conversation_id]
+        else:
+            user_input.conversation_id = conversation_id
+            try:
+                system_message = self._generate_system_message(
+                    exposed_entities, user_input
+                )
+            except TemplateError as err:
+                _LOGGER.error("Error rendering prompt: %s", err)
+                intent_response = intent.IntentResponse(language=user_input.language)
+                intent_response.async_set_error(
+                    intent.IntentResponseErrorCode.UNKNOWN,
+                    f"Sorry, I had a problem with my template: {err}",
+                )
+                return conversation.ConversationResult(
+                    response=intent_response, conversation_id=conversation_id
+                )
+            messages = [system_message]
+        user_message = {"role": "user", "content": user_input.text}
+        if self.entry.options.get(CONF_ATTACH_USERNAME, DEFAULT_ATTACH_USERNAME):
+            user = user_input.context.user_id
+            if user is not None:
+                user_message[ATTR_NAME] = user
+
+        messages.append(user_message)
+
+        try:
+            query_response = await self.query(user_input, messages, exposed_entities, 0)
+        except OpenAIError as err:
+            _LOGGER.error(err)
+            intent_response = intent.IntentResponse(language=user_input.language)
+            intent_response.async_set_error(
+                intent.IntentResponseErrorCode.UNKNOWN,
+                f"Sorry, I had a problem talking to OpenAI: {err}",
+            )
+            return conversation.ConversationResult(
+                response=intent_response, conversation_id=conversation_id
+            )
+        except HomeAssistantError as err:
+            _LOGGER.error(err, exc_info=err)
+            intent_response = intent.IntentResponse(language=user_input.language)
+            intent_response.async_set_error(
+                intent.IntentResponseErrorCode.UNKNOWN,
+                f"Something went wrong: {err}",
+            )
+            return conversation.ConversationResult(
+                response=intent_response, conversation_id=conversation_id
+            )
+
+        messages.append(query_response.message.model_dump(exclude_none=True))
+        self.history[conversation_id] = messages
+
+        self.hass.bus.async_fire(
+            EVENT_CONVERSATION_FINISHED,
+            {
+                "response": query_response.response.model_dump(),
+                "user_input": user_input,
+                "messages": messages,
+            },
+        )
+
+        intent_response = intent.IntentResponse(language=user_input.language)
+        intent_response.async_set_speech(query_response.message.content)
+
+        # Detect if LLM is asking a follow-up question to enable continued conversation
+        response_text = query_response.message.content or ""
+        should_continue = response_text.rstrip().endswith("?") or any(
+            phrase in response_text.lower()
+            for phrase in [
+                "which one",
+                "would you like",
+                "do you want",
+                "would you prefer",
+                "which do you",
+                "what would you",
+                "shall i",
+                "should i",
+                "choose from",
+                "select from",
+                "pick from",
+            ]
+        )
+
+        return conversation.ConversationResult(
+            response=intent_response,
+            conversation_id=conversation_id,
+            continue_conversation=should_continue,
+        )
+
+    def _generate_system_message(
+        self, exposed_entities, user_input: conversation.ConversationInput
+    ):
+        raw_prompt = self.entry.options.get(CONF_PROMPT, DEFAULT_PROMPT)
+        prompt = self._async_generate_prompt(raw_prompt, exposed_entities, user_input)
+        return {"role": "system", "content": prompt}
+
+    def _async_generate_prompt(
+        self,
+        raw_prompt: str,
+        exposed_entities,
+        user_input: conversation.ConversationInput,
+    ) -> str:
+        """Generate a prompt for the user."""
+        return template.Template(raw_prompt, self.hass).async_render(
+            {
+                "ha_name": self.hass.config.location_name,
+                "exposed_entities": exposed_entities,
+                "current_device_id": user_input.device_id,
+            },
+            parse_result=False,
+        )
+
+    def get_exposed_entities(self):
+        states = [
+            state
+            for state in self.hass.states.async_all()
+            if async_should_expose(self.hass, conversation.DOMAIN, state.entity_id)
+        ]
+        entity_registry = er.async_get(self.hass)
+        exposed_entities = []
+        for state in states:
+            entity_id = state.entity_id
+            entity = entity_registry.async_get(entity_id)
+
+            aliases = []
+            if entity and entity.aliases:
+                aliases = entity.aliases
+
+            exposed_entities.append(
+                {
+                    "entity_id": entity_id,
+                    "name": state.name,
+                    "state": self.hass.states.get(entity_id).state,
+                    "aliases": aliases,
+                }
+            )
+        return exposed_entities
+
+    def get_functions(self):
+        try:
+            function = self.entry.options.get(CONF_FUNCTIONS)
+            result = yaml.safe_load(function) if function else DEFAULT_CONF_FUNCTIONS
+            if result:
+                for setting in result:
+                    function_executor = get_function_executor(
+                        setting["function"]["type"]
+                    )
+                    setting["function"] = function_executor.to_arguments(
+                        setting["function"]
+                    )
+            return result
+        except (InvalidFunction, FunctionNotFound) as e:
+            raise e
+        except:
+            raise FunctionLoadFailed()
+
+    async def truncate_message_history(
+        self, messages, exposed_entities, user_input: conversation.ConversationInput
+    ):
+        """Truncate message history."""
+        strategy = self.entry.options.get(
+            CONF_CONTEXT_TRUNCATE_STRATEGY, DEFAULT_CONTEXT_TRUNCATE_STRATEGY
+        )
+
+        if strategy == "clear":
+            last_user_message_index = None
+            for i in reversed(range(len(messages))):
+                if messages[i]["role"] == "user":
+                    last_user_message_index = i
+                    break
+
+            if last_user_message_index is not None:
+                del messages[1:last_user_message_index]
+                # refresh system prompt when all messages are deleted
+                messages[0] = self._generate_system_message(
+                    exposed_entities, user_input
+                )
+
+    async def query(
+        self,
+        user_input: conversation.ConversationInput,
+        messages,
+        exposed_entities,
+        n_requests,
+    ) -> OpenAIQueryResponse:
+        """Process a sentence."""
+        model = self.entry.options.get(CONF_CHAT_MODEL, DEFAULT_CHAT_MODEL)
+        max_tokens = self.entry.options.get(CONF_MAX_TOKENS, DEFAULT_MAX_TOKENS)
+        top_p = self.entry.options.get(CONF_TOP_P, DEFAULT_TOP_P)
+        temperature = self.entry.options.get(CONF_TEMPERATURE, DEFAULT_TEMPERATURE)
+        use_tools = self.entry.options.get(CONF_USE_TOOLS, DEFAULT_USE_TOOLS)
+        context_threshold = self.entry.options.get(
+            CONF_CONTEXT_THRESHOLD, DEFAULT_CONTEXT_THRESHOLD
+        )
+        functions = list(map(lambda s: s["spec"], self.get_functions()))
+        function_call = "auto"
+        if n_requests == self.entry.options.get(
+            CONF_MAX_FUNCTION_CALLS_PER_CONVERSATION,
+            DEFAULT_MAX_FUNCTION_CALLS_PER_CONVERSATION,
+        ):
+            function_call = "none"
+
+        tool_kwargs = {"functions": functions, "function_call": function_call}
+        if use_tools:
+            tool_kwargs = {
+                "tools": [{"type": "function", "function": func} for func in functions],
+                "tool_choice": function_call,
+            }
+
+        if len(functions) == 0:
+            tool_kwargs = {}
+
+        _LOGGER.info("Prompt for %s: %s", model, json.dumps(messages))
+
+        response: ChatCompletion = await self.client.chat.completions.create(
+            model=model,
+            messages=messages,
+            max_tokens=max_tokens,
+            top_p=top_p,
+            temperature=temperature,
+            user=user_input.conversation_id,
+            **tool_kwargs,
+        )
+
+        _LOGGER.info("Response %s", json.dumps(response.model_dump(exclude_none=True)))
+
+        if response.usage and response.usage.total_tokens > context_threshold:
+            await self.truncate_message_history(messages, exposed_entities, user_input)
+
+        choice: Choice = response.choices[0]
+        message = choice.message
+
+        if choice.finish_reason == "function_call" or (
+            choice.finish_reason == "stop" and choice.message.function_call is not None
+        ):
+            return await self.execute_function_call(
+                user_input, messages, message, exposed_entities, n_requests + 1
+            )
+        if choice.finish_reason == "tool_calls" or (
+            choice.finish_reason == "stop" and choice.message.tool_calls is not None
+        ):
+            return await self.execute_tool_calls(
+                user_input, messages, message, exposed_entities, n_requests + 1
+            )
+        if choice.finish_reason == "length":
+            raise TokenLengthExceededError(response.usage.completion_tokens)
+
+        return OpenAIQueryResponse(response=response, message=message)
+
+    async def execute_function_call(
+        self,
+        user_input: conversation.ConversationInput,
+        messages,
+        message: ChatCompletionMessage,
+        exposed_entities,
+        n_requests,
+    ) -> OpenAIQueryResponse:
+        function_name = message.function_call.name
+        function = next(
+            (s for s in self.get_functions() if s["spec"]["name"] == function_name),
+            None,
+        )
+        if function is not None:
+            return await self.execute_function(
+                user_input,
+                messages,
+                message,
+                exposed_entities,
+                n_requests,
+                function,
+            )
+        raise FunctionNotFound(function_name)
+
+    async def execute_function(
+        self,
+        user_input: conversation.ConversationInput,
+        messages,
+        message: ChatCompletionMessage,
+        exposed_entities,
+        n_requests,
+        function,
+    ) -> OpenAIQueryResponse:
+        function_executor = get_function_executor(function["function"]["type"])
+
+        try:
+            arguments = json.loads(message.function_call.arguments)
+        except json.decoder.JSONDecodeError as err:
+            raise ParseArgumentsFailed(message.function_call.arguments) from err
+
+        result = await function_executor.execute(
+            self.hass, function["function"], arguments, user_input, exposed_entities
+        )
+
+        messages.append(
+            {
+                "role": "function",
+                "name": message.function_call.name,
+                "content": str(result),
+            }
+        )
+        return await self.query(user_input, messages, exposed_entities, n_requests)
+
+    async def execute_tool_calls(
+        self,
+        user_input: conversation.ConversationInput,
+        messages,
+        message: ChatCompletionMessage,
+        exposed_entities,
+        n_requests,
+    ) -> OpenAIQueryResponse:
+        messages.append(message.model_dump(exclude_none=True))
+        for tool in message.tool_calls:
+            function_name = tool.function.name
+            function = next(
+                (s for s in self.get_functions() if s["spec"]["name"] == function_name),
+                None,
+            )
+            if function is not None:
+                result = await self.execute_tool_function(
+                    user_input,
+                    tool,
+                    exposed_entities,
+                    function,
+                )
+
+                messages.append(
+                    {
+                        "tool_call_id": tool.id,
+                        "role": "tool",
+                        "name": function_name,
+                        "content": str(result),
+                    }
+                )
+            else:
+                raise FunctionNotFound(function_name)
+        return await self.query(user_input, messages, exposed_entities, n_requests)
+
+    async def execute_tool_function(
+        self,
+        user_input: conversation.ConversationInput,
+        tool,
+        exposed_entities,
+        function,
+    ) -> OpenAIQueryResponse:
+        function_executor = get_function_executor(function["function"]["type"])
+
+        try:
+            arguments = json.loads(tool.function.arguments)
+        except json.decoder.JSONDecodeError as err:
+            raise ParseArgumentsFailed(tool.function.arguments) from err
+
+        result = await function_executor.execute(
+            self.hass, function["function"], arguments, user_input, exposed_entities
+        )
+        return result
+
+
+class OpenAIQueryResponse:
+    """OpenAI query response value object."""
+
+    def __init__(
+        self, response: ChatCompletion, message: ChatCompletionMessage
+    ) -> None:
+        """Initialize OpenAI query response value object."""
+        self.response = response
+        self.message = message
